cors habilitado

+++ controllers/peliculasControllers.js
@@ -1,14 +1,16 @@
 const asyncHandler = require('express-async-handler')
 const Pelicula = require('../model/peliculaModel')
 
+
 const getPeliculas = asyncHandler(async (req, res) => {
-    const peliculas = await Pelicula.find({user: req.user.id})
+    const peliculas = await Pelicula.find({ user: req.user.id })
     res.status(200).json(peliculas)
 })
 
-const createPeliculas = asyncHandler(async(req, res) => {
-    if(!req.body){
-        res.status(400).json({message: "escribe"})
+
+const createPeliculas = asyncHandler(async (req, res) => {
+    if (!req.body) {
+        res.status(400).json({ message: "escribe" })
         throw new Error('Escribe el titulo de la pelilcula')
     }
     const pelicula = await Pelicula.create({
@@ -28,22 +30,30 @@ const createPeliculas = asyncHandler(async(req, res) => {
         vote_count: req.body.vote_count,
         user: req.user.id
     })
-        
+
     res.status(201).json(pelicula)
 })
 
+
 const updatePelicula = asyncHandler(async (req, res) => {
-    
     const pelicula = await Pelicula.findById(req.params.id)
 
     if (!pelicula) {
         res.status(400)
-        throw new Error ('Pelicula no encontrada')
-    }else{
-        const peliculaUpdated = await Pelicula.findByIdAndUpdate(req.params.id, req.body, {new: true}) //Devuelve elemento ya modificado
-        res.status(200).json(peliculaUpdated)
+        throw new Error('Pelicula no encontrada')
+    } else {
+
+        //verificacion de pertenencia al usuario
+        if (pelicula.user.toString() !== req.user.id) {
+            res.status(401)
+            throw new Error('Usuario no autorizado')
+        } else {
+            const peliculaUpdated = await Pelicula.findByIdAndUpdate(req.params.id, req.body, { new: true }) //Devuelve elemento ya modificado
+            res.status(200).json(peliculaUpdated)
+        }
     }
 })
+
 
 const deletePelicula = asyncHandler(async (req, res) => {
     const pelicula = await Pelicula.findById(req.params.id)
@@ -51,9 +61,15 @@ const deletePelicula = asyncHandler(async (req, res) => {
     if (!pelicula) {
         res.status(400)
         throw new Error('Pelicula no encontrada')
-    }else{
-        await pelicula.deleteOne()
-        res.status(200).json({id: req.params.id})
+    } else {
+        //verificacion de pertenencia al usuario
+        if (pelicula.user.toString() !== req.user.id) {
+            res.status(401)
+            throw new Error('Usuario no autorizado')
+        } else {
+            await pelicula.deleteOne()
+            res.status(200).json({ id: req.params.id })
+        }
     }
 })
